+++ NEWS.md
@@ -25,6 +25,9 @@
 -   Updated sparse matrix conversion routines for compatibility with
     scipy 1.11.0.
 
+-   Fixed an issue where a py capsule finalizer could access the R API from
+    a background thread. (#1406)
+
 # reticulate 1.30
 
 - Fix compilation error on R 3.5. Bump minimum R version dependency to 3.5.